refactor(next/web): TicketOverview
